requirements.txt
flask
flask-assistant
requests
python-docx
wakeonlan
aniso8601
